MDL-57755 MOODLE_34_STABLE CODE AREA: messaging and notifications
As far as I am aware Moodle had messages and notifications functionality together and there was some sort of taking the
functionalities away from each other but apparently it's not done yet. We have been reported a bug that users dont see new
notifications and that the notifications bell is hidden. On our investigations we found that for some reason all the
notifications are marked as read as soon as they come to life. For a first resolution we suggested our team to turn on
messaging as that would 'fix' the notifications. We have prioritized it again and as I continued working on it I found
that anything thats not a message (but treated like a message ==> notifications) are delt with based on $CFG->messaging.
 This means that if you decide not to use Instant messaging on your Moodle instance, that would break notifications.
I've been advised to create a new functionality that is as simple as possible to update all the $CFG->messaging
functionality when it's used as a notification.

To fix this I ADDED the following functionality:
/admin/settings/subsystems.php:18
new admin settings to turn on/off Notifications ( similar to messaging )

/lang/en/admin.php:289
language strings for the new admin settings above.

/lib/classes/message/manager.php:165
if this is a notification and notifications are turned off, mark it read.
with the fix on line 159, we would still have notifications unread, so if later
Notifications turned back ON, users could have 10's or 100's of unread
notifications. This condition fixes that.

And UPDATED the following:
/lib/classes/message/manager.php:159
only mark the message(notification) read if Notifications are turned off this is
the actual piece of code that allowed us to have notification unread if messaging
is turned off.

/message/externallib.php:1854
So, our notifications now appear as unread while Messaging turned off, but we still
had a condition here which throwed error when clicked View full notification or See all
because the notification page was using the same page as messages and was only checking
for $CFG->messaging. Users also wouldn't be able to mark notifications as read.

/message/output/popup/lib.php:63
This update is to only show the notifications bell if the Notifications are turned on